% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinhole.R
\name{RT.PinholeCamera}
\alias{RT.PinholeCamera}
\title{Constructs a valid pinhole camera object. In particular it defines the notional internal dimensions of said camera, and its location in **vecspace**.}
\usage{
RT.PinholeCamera(hole.location = c(0, 0, 0), film.centre = c(0, 0, -0.052),
  film.up = c(0, 1, 0), film.right = c(1, 0, 0), film.size = 0.036)
}
\arguments{
\item{hole.location}{The notional position of the pinhole in x,y,z space. Defaults to c(0,0,0)}

\item{film.centre}{The notional location of the centre of the film within the camera. The default properties of all of the film.parameters
  define a film located back along the negative z axis (i.e "looking" in the (0,0,1) direction) with a size adn subsequent view angle
  approximating that of a ~50mm SLR lens.
@param film.up The notional orientation of the film. By default is "up" i.e to take a horizontal aligned picture (where the positive y axis defines up)
@param film.right Further property of the notional orientation of the film. By default is orthogonal to the view direction and the up vector,
  so as to define a rectangular film orthogonal to the view direction. Interesting photograph techniques could be created by manipulating these
  film values to align the film in a different orientation to the view direction.
@param film.size The notional size of the film to be exposed. This defines a square size, which by default approximates a 35mm film.
  The aspect ratio of the actual exposure will be defined by image height and width (in pixels) with *film.size* taking the longer of the two
  pixel dimensions.}
}
\value{
A validated camera object.
}
\description{
Constructs a valid pinhole camera object. In particular it defines the notional internal dimensions of said camera, and its location in **vecspace**.
}
\examples{
  cam <- RT.PinholeCamera()
}
