% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface.R
\name{RT.Surface}
\alias{RT.Surface}
\title{Validate parameters, and return a valid surface object}
\usage{
RT.Surface(rgb, reflectivity = 1, reflectivity.matte = 0,
  transparency = 0, transparency.matte = 0)
}
\value{
Surface object
}
\description{
Validate parameters, and return a valid surface object
}
\examples{
  surf <- RT.asSurface(rgb=c(1,1,0)))
}
