% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface.R
\name{RT.Surface}
\alias{RT.Surface}
\title{Validate parameters, and return a valid surface object}
\usage{
RT.Surface(rgb, matteness = 0, transparency = 0)
}
\arguments{
\item{rgb}{Either a 3 value vector of numbers in teh range 0 to 1, or a function which will return the equivalent.
The function, if present, will be passed the values of *north* and *east* as returned by the object's intersection.
The function allow interesting surface properties to be easily constructed, for example patterned, image mapping, smoothing or 
randomisation.}

\item{matteness}{The proportion of matte reflectivity of the surface. 1 minus this value will be reflected as mirror reflection.
Default = 0, i.e the surface has mirror reflectivity}

\item{transparency}{The proportion of incident light which will be refracted by the surface. 1 minus this value will be reflected.
Default = 0, i.e. a non-transparent surface.}
}
\value{
A validated surface object. Ray tracing functions in this package expect all solid objects to have a valid surface property.
}
\description{
Validate parameters, and return a valid surface object
}
\examples{
  surf <- RT.Surface(rgb=c(1,1,0), transparency=1)
}
